fix: move routine_storage_walk tests out of an inline #[cfg(test)] block
The repo's own pre-push hook (and CONTRIBUTING.md) forbid inline
`#[cfg(test)] mod tests { ... }` blocks in favor of *_tests.rs sibling
files, but src/routine_storage_walk.rs had one. No CI job mirrors this
specific check, so it silently broke `main` for anyone actually running
`.githooks/pre-push`. Relocate the two tests to
src/routine_storage_walk_tests.rs, matching the convention used
everywhere else in the crate.
